+++ android/src/org/coolreader/crengine/ReaderOptionsHandler.java
@@ -1,0 +1,24 @@
+/*
+ * CoolReader for Android
+ * Copyright (C) 2026 CoolReader Next contributors
+ *
+ * This program is free software: you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation, either version 2
+ * of the License, or (at your option) any later version.
+ */
+
+package org.coolreader.crengine;
+
+interface ReaderOptionsHandler {
+	boolean isActive();
+
+	ReaderDocumentOptions snapshot();
+
+	boolean applyDocumentOptions(
+			boolean textAutoformatEnabled,
+			boolean documentStylesEnabled,
+			boolean documentFontsEnabled,
+			int domVersion,
+			int blockRenderingFlags);
+}